# internal/docs/openapi.yaml
openapi: 3.0.0
info:
  title: feedsanctum
  description: RSS feed aggregator
  version: 1.0.0
servers:
  - url: /
    description: Production render
  - url: http://localhost:8081
    description: Local development
...
paths:
...
  /api/v1/feeds:
    get:
      summary: Retrieves a list of all registered RSS feeds in the system.
      operationId: listFeeds
      tags:
        - Feeds
      responses:
        '200':
          description: Returned with a JSON array of feeds. Returns an empty array if no feeds exist.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Feed'
              examples:
                feedsList:
                  summary: List with two feeds
                  value:
                    - id: "e4b01140-5b5c-4f9e-9764-77a83d7cb45d"
                      url: "https://exampletech.com/index.xml"
                      category: "tech"
                      last_fetched_at: "2025-01-15T10:30:00Z"
                      created_at: "2025-01-14T08:00:00Z"
                      updated_at: "2025-01-15T10:30:00Z"
                    - id: "76fa2e7d-304e-4e44-b054-945fa5cb9c20"
                      url: "https://news.ycombinator.com/rss"
                      category: "news"
                      last_fetched_at: null
                      created_at: "2025-01-15T09:00:00Z"
                      updated_at: "2025-01-15T09:00:00Z"
                emptyList:
                  summary: No feeds registered
                  value: []
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          description: Returned when a database error occurs while fetching feeds.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                databaseFailure:
                  summary: Database failure
                  value:
                    error: "something went wrong" 

    post:
      summary: Registers a new RSS feed in the system within a transaction and automatically creates a feed follow entry with a custom label for the authenticated user.
      operationId: createFeed
      tags:
        - Feeds
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFeedRequest'
            examples:
              createFeed:
                summary: Register a tech blog feed
                value:
                  url: "https://exampletech.com/index.xml"
                  name: "My Tech Blog"
      responses:
        '201':
          description: Returned when the feed is successfully created and followed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedFollow'
              examples:
                feedCreated:
                  summary: Feed created and followed
                  value:
                    id: "c1f76e10-90fb-4d89-9a06-4b8c6e2467b2"
                    user_id: "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11"
                    feed_id: "e4b01140-5b5c-4f9e-9764-77a83d7cb45d"
                    name: "My Tech Blog"
                    user_name: "Jane Doe"
                    created_at: "2025-01-15T12:00:00Z"
                    updated_at: "2025-01-15T12:00:00Z"
        '400':
          description: Returned when the request body is malformed or input validation fails.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                malformedJson:
                  summary: Malformed JSON payload
                  value:
                    error: "invalid request body"
                invalidUrl:
                  summary: Invalid feed URL
                  value:
                    error: "invalid URL"
                invalidName:
                  summary: Invalid feed name
                  value:
                    error: "invalid feed name"
        '401':
          description: Returned when the access token is missing or invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                authTokenInvalid:
                  summary: Token missing or invalid
                  value:
                    error: "invalid access token"
        '409':
          description: Returned when the feed URL is already registered in the system.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                duplicateFeed:
                  summary: Feed URL already registered
                  value:
                    error: "feed already registered"
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          description: Returned when context lookup fails or a database transaction error occurs.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                databaseFailure:
                  summary: Database or transaction failure
                  value:
                    error: "something went wrong"
                missingContext:
                  summary: Missing user ID in context
                  value:
                    error: "missing user id in context"           
# --------------------------------------------------------------- 
  /api/v1/feeds/{feedID}:
    get:
      summary: Get feed by ID
      operationId: getFeedById
      tags:
        - Feeds
      parameters:
        - name: feedID
          in: path
          required: true
          description: The UUID of the feed to retrieve.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Returned when the feed is found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Feed'
              examples:
                feedFound:
                  summary: Feed found
                  value:
                    id: "e4b01140-5b5c-4f9e-9764-77a83d7cb45d"
                    url: "https://exampletech.com/index.xml"
                    category: "tech"
                    last_fetched_at: "2025-01-15T10:30:00Z"
                    created_at: "2025-01-14T08:00:00Z"
                    updated_at: "2025-01-15T10:30:00Z"
        '400':
          description: Returned when the provided feedID is not a valid UUID format.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidFeedID:
                  summary: Invalid feed ID format
                  value:
                    error: "invalid feed id"
        '404':
          description: Returned when no feed matches the provided feedID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                feedNotFound:
                  summary: Feed not found
                  value:
                    error: "no such feed found"
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          description: Returned when a database error occurs while fetching the feed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                databaseFailure:
                  summary: Database failure
                  value:
                    error: "something went wrong"
# --------------------------------------------------------------- 
...
  /api/v1/follows/{feedID}:
    delete:
      summary: Unfollows a feed for the authenticated user.
      operationId: deleteFeedFollow
      tags:
        - Feeds
      security:
        - bearerAuth: []
      parameters:
        - name: feedID
          in: path
          required: true
          description: The UUID of the feed to unfollow.
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Returned when the feed follow is successfully deleted. No response body is returned.
        '400':
          description: Returned when the provided feedID in the URL is not a valid UUID format.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidFeedID:
                  summary: Invalid feed ID format
                  value:
                    error: "invalid feed id"
        '401':
          description: Returned when the access token is missing or invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                authTokenInvalid:
                  summary: Token missing or invalid
                  value:
                    error: "invalid access token"
        '404':
          description: Returned when the authenticated user is not following the specified feed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                feedFollowNotFound:
                  summary: Feed follow not found
                  value:
                    error: "no such feed follow found"
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          description: Returned when context lookup fails or a database error occurs while deleting the follow.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                databaseFailure:
                  summary: Database failure
                  value:
                    error: "something went wrong"
                missingContext:
                  summary: Missing user ID in context
                  value:
                    error: "missing user id in context"
# --------------------------------------------------------------- 
...
components:
  schemas:
...
    Feed:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        category:
          type: string
        last_fetched_at:
          type: string
          format: date-time
          nullable: true
          description: Null if the feed has never been fetched.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - url
        - category
        - created_at
        - updated_at
        - last_fetched_at
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    CreateFeedRequest:
      type: object
      properties:
        url:
          type: string
          format: uri
          description: The RSS feed URL.
        name:
          type: string
          minLength: 1
          maxLength: 45
          description: Custom name/label for the feed follow.
      required:
        - url
        - name
    FeedFollow:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        feed_id:
          type: string
          format: uuid
        name:
          type: string
          description: The user's custom label for this feed.
        user_name:
          type: string
          description: The display name of the user who follows this feed.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - user_id
        - feed_id
        - name
        - user_name
        - created_at
        - updated_at
...
  responses:
      RateLimitExceeded:
        description: Returned when the client exceeds the allowed request rate.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ErrorResponse'
            examples:
              ipLimitExceeded:
                summary: Too many requests from this IP
                value:
                  error: "Rate limit reached"
              loginAttemptsExceeded:
                summary: Too many login attempts for this email
                value:
                  error: "login attempts exceeded, please try again later"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token for authenticated endpoints
    refreshTokenAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Refresh token for /refresh and /logout
security: [] # public